refac: move utility functions
